flask-api/requirements.txt
tensorflow==2.10
flask
numpy
pillow
opencv-python

